Create the Connect-5 playing environment [Update 7]
This commit creates the board which the user will play with for their game of Connect5.
One of the most essential aspects of this commit is the is_valid function. is_valid  is used to detect whether or not the player is trying to place their piece within the dimensions of the game board. This is important because the game pieces have to be placed on the game’s grid while the player is trying to achieve connecting 5 pieces in a row.

Remove unnecessary for-loops [Update 6]

This commit was done to remove unnecessary for-loops in several of the functions. This commit is necessary as our group would like out code as clean, concise, and elegant as possible. This is also for the sake of make the code more efficient.

Complete set_piece [Update 5]

This commit was done to establish the completion of set_piece. set_piece is a necessary part of this project as the game board needs to be updated whenever a new piece is added to the grid.

Complete get_piece [Update 4]

This commit was done to complete the get_piece function. get_piece is used to help determine whether the spot is empty, occupied by player1, or is occupied by player2. This commit is important because this information will be used when determining if a player has won the game in the function has_connect5.

Struggle importing the Player and Game classes [Update 3]

This commit was done as an attempt to import the Player class from Player.py, and the Game class from Game.py. However, the imports were not achievable.

Create the grid [Update 2]

This commit was done to create a nested for-loop under __init__ create an empty dimension x dimension grid through the means of a nested list. This was necessary as the information in regards to the where the pieces are placed on the board will be kept track of using this list of lists.

Specify the data type [Update 1]

This commit was done to specify the data type of the variable dimensions in __init__ as it wasn't specified earlier.