Check the grid input of the user
This commit is done to establish the completion of the is_valid function which is essential to the board class. is_valid  is used to detect whether or not the player is trying to place their piece within the dimensions of the game board. This is important because the game pieces have to be placed on the game’s grid in order for the player to win when they have five of their own pieces in a line.