Complete the has_connect5 and has_connect4 function
This commit is done to establish the completion of the has_connect5 function which is essential to the board class. The function has_connect5  is used to detect whether or not a player has won the game. This is important because the entire game of connect5 is completed when there are 5 consecutive pieces forming a linear line. This is also done to help declare when there is a winner.

The function has_connect4 is used to help make an intermediate level AI block off winning moves of the user. This concept is used to set part the easy leveled AI from the intermediate levelled AI. As a result, the function has_connect4 is necessary for this project.